+++ api/migrations/0016_little_valkyrie.sql
@@ -1,0 +1,6 @@
+ALTER TABLE "tests" ADD COLUMN "assignee_id" uuid;--> statement-breakpoint
+DO $$ BEGIN
+ ALTER TABLE "tests" ADD CONSTRAINT "tests_assignee_id_users_id_fk" FOREIGN KEY ("assignee_id") REFERENCES "public"."users"("id") ON DELETE set null ON UPDATE no action;
+EXCEPTION
+ WHEN duplicate_object THEN null;
+END $$;